minor cleanups and rename helper files to align with DSLs
