+++ .changeset/big-squids-lick.md
@@ -1,0 +1,9 @@
+---
+'@finos/legend-extension-dsl-diagram': patch
+'@finos/legend-extension-external-store-service': patch
+'@finos/legend-extension-mapping-generation': patch
+'@finos/legend-graph': patch
+'@finos/legend-query': patch
+'@finos/legend-studio': patch
+'@finos/legend-studio-extension-query-builder': patch
+---
