+++ packages/legend-graph/src/index.ts
@@ -157,9 +157,9 @@ export { TDSResultType } from './models/metamodels/pure/executionPlan/result/TDS
 export * from './MetaModelUtils';
 export * from './MetaModelConst';
 
-export * from './helpers/ServiceHelper';
-export * from './helpers/DatabaseHelper';
-export * from './helpers/MappingHelper';
+export * from './helpers/DSLService_Helper';
+export * from './helpers/StoreRelational_Helper';
+export * from './helpers/DSLMapping_Helper';
 export * from './helpers/ValidationHelper';
 export * from './helpers/DomainHelper';
 export * from './helpers/PureLanguageHelper';
