+++ packages/legend-graph/src/models/metamodels/pure/packageableElements/generationSpecification/GenerationSpecification.ts
@@ -22,10 +22,10 @@ import { type Hashable, hashArray } from '@finos/legend-shared';
 import {
   type PackageableElementVisitor,
   PackageableElement,
-  getElementPointerHashCode,
 } from '../PackageableElement';
 import type { PackageableElementReference } from '../PackageableElementReference';
 import type { FileGenerationSpecification } from '../fileGeneration/FileGenerationSpecification';
+import { hashElementPointer } from '../../../../../MetaModelUtils';
 
 // NOTE: As of now the tree only supports a linear order of generation. This is because the only use case is linear,
 // but the shape has been left as a tree to support 'branching' off in the future.
@@ -70,7 +70,7 @@ export class GenerationSpecification
       hashArray(this.generationNodes),
       hashArray(
         this.fileGenerations.map((fileGeneration) =>
-          getElementPointerHashCode(
+          hashElementPointer(
             PackageableElementPointerType.FILE_GENERATION,
             fileGeneration.hashValue,
           ),
