+++ packages/legend-graph/src/models/protocols/pure/v1/transformation/pureGraph/to/V1_ProtocolToMetaModelClassMappingSecondPassBuilder.ts
@@ -64,7 +64,7 @@ import {
   getAllClassMappings,
   getAllEnumerationMappings,
   getOwnClassMappingById,
-} from '../../../../../../../helpers/MappingHelper';
+} from '../../../../../../../helpers/DSLMapping_Helper';
 import type { DSLMapping_PureProtocolProcessorPlugin_Extension } from '../../../../DSLMapping_PureProtocolProcessorPlugin_Extension';
 import type { V1_MergeOperationClassMapping } from '../../../model/packageableElements/mapping/V1_MergeOperationClassMapping';
