+++ packages/legend-graph/src/models/protocols/pure/v1/transformation/pureGraph/to/helpers/V1_AssociationMappingHelper.ts
@@ -27,7 +27,7 @@ import { V1_ProtocolToMetaModelPropertyMappingBuilder } from '../../../../transf
 import type { V1_AssociationMapping } from '../../../../model/packageableElements/mapping/V1_AssociationMapping';
 import { V1_XStoreAssociationMapping } from '../../../../model/packageableElements/mapping/xStore/V1_XStoreAssociationMapping';
 import { XStoreAssociationImplementation } from '../../../../../../../metamodels/pure/packageableElements/mapping/xStore/XStoreAssociationImplementation';
-import { getAllEnumerationMappings } from '../../../../../../../../helpers/MappingHelper';
+import { getAllEnumerationMappings } from '../../../../../../../../helpers/DSLMapping_Helper';
 
 const getInferredAssociationMappingId = (
   _association: Association,
